backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.1
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
numpy==1.26.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiosmtplib==3.0.1
jinja2==3.1.2
email-validator==2.1.0
aiofiles==23.2.1
httpx==0.25.2
